requirements.txt
astroid==2.4.2
click==7.1.2
Flask==1.1.2
gunicorn==20.0.4
isort==5.7.0
itsdangerous==1.1.0
Jinja2==2.11.2
lazy-object-proxy==1.4.3
MarkupSafe==1.1.1
mccabe==0.6.1
pylint==2.6.0
six==1.15.0
toml==0.10.2
Werkzeug==1.0.1
wrapt==1.12.1
google-api-core==2.5.0
google-api-python-client==2.38.0
google-auth==2.6.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.5.0
googleapis-common-protos==1.55.0